requirements.txt
alabaster==0.7.11
amqp==2.3.2
appnope==0.1.0
argh==0.26.2
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
cairocffi==0.9.0
CairoSVG==2.1.3
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
CommonMark==0.5.4
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.8
django-bootstrap3==11.0.0
django-bootstrap3-datetimepicker-3==2.6.0
django-ckeditor==5.6.1
django-generate-secret-key==1.0.2
django-js-asset==1.1.0
django-pure-pagination==0.3.0
docutils==0.14
enum34==1.1.6
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.7
imagesize==1.1.0
ipython==6.5.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
kombu==4.2.1
livereload==2.5.2
MarkupSafe==1.0
openpyxl==2.5.6
packaging==17.1
parso==0.3.1
pathlib2==2.3.2
pathtools==0.1.2
pdfrw==0.4
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
port-for==0.3.1
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
Pyphen==0.9.5
pytz==2018.5
PyYAML==3.13
recommonmark==0.4.0
redis==2.10.6
requests==2.19.1
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.8
sphinx-autobuild==0.7.1
sphinxcontrib-websupport==1.1.0
tinycss2==0.6.1
tornado==5.1
traitlets==4.3.2
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.4
watchdog==0.9.0
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
